Renamed location.php to Location.php
Changed this directly on Github, because (apparently) OSX is case insensitive when it comes to filenames. So Git wasn't picking up any changes when I altered the filename. Also PSR-4 is case sensitive, meaning composer couldn't find the Location class because it was looking for a file in /src/ called location.php (note the lowercase l). Since Travis is running on Linux and Linux is case sensitive when it comes to filenames, I'm hoping this will pass the Travis tests.